refactor: Genre
- structured new folder for Genre and related files
- added new animations for struct GenreChipView + GenreSelectorView
